requirements.txt
Flask==2.3.3
gunicorn==20.1.0
numpy==1.26.4
Pillow==10.3.0
tensorflow-cpu==2.16.1